Fix an infinite loop in ErlangDist::setParams
This method was trying to call the superclass (GammaDist) version that takes a double
argument for the k value.  Add a cast to force it to call the correct method instead
of calling itself.  Could also have been fixed by adding an explicit super to the
call.

+++ source/umontreal/iro/lecuyer/probdist/ErlangDist.java
@@ -207,7 +207,7 @@ public class ErlangDist extends GammaDist {
     * 
     */
    public void setParams (int k, double lambda, int d) {
-      setParams (k, lambda, d);
+      setParams ((double)k, lambda, d);
    }
